# WIP: Navigator with max constrained step size under Arbitray B field and Interactions 