--- a/core/include/detray/propagator/navigator.hpp
+++ b/core/include/detray/propagator/navigator.hpp
@@ -751,7 +751,6 @@ class navigator {
                                              navigation::void_inspector>) {
                 navigation.run_inspector("Update complete: fair trust: ");
             }
-
             return;
         }