Motivation: to make navigator work with max constrained step size and arbitrary b field and material interaction

There are two edge cases that I was trying to handle in the high trust case of `update_kernel` in navigator
1. the track propagated to the next surface oversteps and misses the target surface
-> revert the track of stepper to the previous state and redo the stepping with the half step size
2. the actual track never meets the target surface (Not sure the algorithm in the PR is perfect)
->  Change the trust level to `no_trust` to get the new candidates

I had to cache the previous track state and move the overstep tolerance into the stepper. Eventually I will need to remove the overstep tolerance in free track parameter and ray

It seemed working somehow before #479 - for example, the last test of `tools_propagator.cpp`  did work with max constrained step size and the run time is reduced by half. (Only one track of the second test fails which I don't know why but I think that can be fixed later.)

```
before
[----------] Global test environment set-up.
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0 (285 ms)
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper (285 ms total)

[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0 (973 ms)
[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper (973 ms total)

after
[----------] Global test environment set-up.
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0 (292 ms)
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper (292 ms total)


[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0 (436 ms)
[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper (436 ms total)

```

After rebasing with #479, `tools_propagator.cpp` breaks because the search function of volume grid yields a segmentation fault :disappointed:  

The PR itself might change a lot after discussion but I think the current navigator should be improved to make it work with the max constrained step size to save the time