@beomki-yeo 2023-06-09
Another issue is that `volume_by_pos` seems not compiled with cuda
```
/home/beomki/projects/detray/detray/core/include/detray/core/detector.hpp(260): warning #20011-D: calling a __host__ function("auto detray::grid< ::detray::n_axis::multi_axis<(bool)1,  ::detray::cylindrical3< ::algebra::cmath::transform3< ::algebra::cmath::matrix::actor<unsigned long,  ::algebra::matrix::array_type,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::matrix::determinant::actor<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::matrix::determinant::partial_pivot_lud<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float>   > ,  ::algebra::cmath::matrix::determinant::hard_coded<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float> , (unsigned long)2ul, (unsigned long)4ul >  > ,  ::algebra::cmath::matrix::inverse::actor<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::matrix::inverse::partial_pivot_lud<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float>   > ,  ::algebra::cmath::matrix::inverse::hard_coded<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float> , (unsigned long)2ul, (unsigned long)4ul >  > ,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float> ,  ::algebra::cmath::block_getter<unsigned long,  ::algebra::matrix::array_type, float> > > > ,  ::detray::n_axis::single_axis< ::detray::n_axis::open<( ::detray::n_axis::label)0> ,  ::detray::n_axis::irregular< ::detray::container_types< ::vecmem::device_vector,  ::detray::tuple,  ::detray::darray,  ::vecmem::jagged_device_vector,  ::detray::dmap> , float> > ,  ::detray::n_axis::single_axis< ::detray::n_axis::circular<( ::detray::n_axis::label)1> ,  ::detray::n_axis::regular< ::detray::container_types< ::vecmem::device_vector,  ::detray::tuple,  ::detray::darray,  ::vecmem::jagged_device_vector,  ::detray::dmap> , float> > ,  ::detray::n_axis::single_axis< ::detray::n_axis::open<( ::detray::n_axis::label)2> ,  ::detray::n_axis::irregular< ::detray::container_types< ::vecmem::device_vector,  ::detray::tuple,  ::detray::darray,  ::vecmem::jagged_device_vector,  ::detray::dmap> , float> >  > , unsigned int,  ::detray::simple_serializer,  ::detray::replacer> ::global_to_local< ::algebra::cmath::transform3< ::algebra::cmath::matrix::actor<unsigned long,  ::algebra::matrix::array_type,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::matrix::determinant::actor<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::matrix::determinant::partial_pivot_lud<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float>   > ,  ::algebra::cmath::matrix::determinant::hard_coded<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float> , (unsigned long)2ul, (unsigned long)4ul >  > ,  ::algebra::cmath::matrix::inverse::actor<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::matrix::inverse::partial_pivot_lud<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float>   > ,  ::algebra::cmath::matrix::inverse::hard_coded<unsigned long,  ::algebra::matrix::matrix_type, float,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float> , (unsigned long)2ul, (unsigned long)4ul >  > ,  ::algebra::cmath::element_getter<unsigned long,  ::algebra::matrix::array_type, float> ,  ::algebra::cmath::block_getter<unsigned long,  ::algebra::matrix::array_type, float> > > ,     ::std::array<float, (unsigned long)3ul> ,     ::std::array<float, (unsigned long)3ul> > (const T1 &, const T2 &, const T3 &) const") from a __host__ __device__ function("detray::detector< ::detray::toy_metadata< ::covfie::backend::constant< ::covfie::vector::vector_d<float, (unsigned long)3ul> ,  ::covfie::vector::vector_d<float, (unsigned long)3ul> > > ,  ::covfie::field_view,  ::detray::container_types< ::vecmem::device_vector,  ::detray::tuple,  ::detray::darray,  ::vecmem::jagged_device_vector,  ::detray::dmap> > ::volume_by_pos const") is not allowed
```