@niermann999 2023-06-10
> There are two edge cases that I was trying to handle in the high trust case of `update_kernel` in navigator

Why do you add this only to the high trust case? The cases I could imagine right now when the normal overstepping handling might not work also happens in all other trust cases or not? Especially the query to the volume grid needs to go into the init funciton once it actually works.

> I had to cache the previous track state and move the overstep tolerance into the stepper.

I am skeptical about this, since the previous track state is huge and almost never needed, so most of the time this memory is wasted (By the way, do we know how often the normal overstepping correction fails?). But maybe the propagation state is already so large it does not make much of a difference. What is the impact of this on the (device) propagation benchmark?

> Eventually I will need to remove the overstep tolerance in free track parameter and ray

I don't think it is a good idea to move the overstepping tolerance to the stepper state, since it is also used to make the intersectors more efficient and I don't think we should pass the stepper state there. The track state is part of the stepper state, so the overstepping tolerance on the track should already be available there or not?