@beomki-yeo 2023-06-13
@niermann999 please don't worry too much. I will not push this further when it turns out to be not beneficial for CKF. I will make a reasonable benchmark before asking for a serious review.

I want to remove the overstepping tolerance in free track parameter for one good reason: we don't need a free covariance matrix in stepper or navigator which is taking large amount of memory. What we need is free vector and overstep tolerance, which means that we don't have to couple vector, matrix, tolerance in a single class type.

I think the tolerance could be put in a more reasonable place, e.g. navigator or stepper.

> Why do you add this only to the high trust case?

We definitely don't have to do this for full and no_trust right? I need to think about what to do for fair trust but I am not sure at this stage

> do we know how often the normal overstepping correction fails?

I think many times in oddly aligned b field