@niermann999 2023-06-13
Since we don't have a meeting today let me post this here:
Setting a larger overstepping tolerance and less strict stepsize constraints also results in a similar speedup. 

I believe, that at least part of the reason the propagation is failing here, is that the toy geometry is not designed for a bfield that is not in z-direction. The beampipe volume is very long in z, because the assumption is, that all tracks traveling in that direction are essentially straight lines and it is thus safe to transport them over large distances. If the bfield does not point in z, then the stepper will still try to step all the way at once, but since the tracks now bend into the detector, the final position will be starting to leave the beampipe. 

Preventing the stepper from making large steps when the track direction changes significantly over the stepping distance seems to solve the problem for all three test cases and reduces the unittest execution time even more. That does not mean it working correclty though, as surface may still be skipped altogether.
The flipside is, that this is still solving the problem by taking more steps, which is costly in terms of execution time and hence a bad solution for low momentum tracks. Also, the thing I hacked together here is still super stupid :D and needs a better implementation 

before
```
Note: Google Test filter = PropagatorValidation*
[==========] Running 4 tests from 4 test suites.
[----------] Global test environment set-up.
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0 (249 ms)
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper (249 ms total)

[----------] 1 test from PropagatorValidation2/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation2/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation2/PropagatorWithRkStepper.propagator_rk_stepper/0 (895 ms)
[----------] 1 test from PropagatorValidation2/PropagatorWithRkStepper (896 ms total)

[----------] 1 test from PropagatorValidation3/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation3/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation3/PropagatorWithRkStepper.propagator_rk_stepper/0 (896 ms)
[----------] 1 test from PropagatorValidation3/PropagatorWithRkStepper (896 ms total)

[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0 (898 ms)
[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper (898 ms total)

[----------] Global test environment tear-down
[==========] 4 tests from 4 test suites ran. (2939 ms total)
[  PASSED  ] 4 tests.

 Running ./bin/detray_array_propagator_cuda
Run on (48 X 1793.77 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 1.32, 3.92, 2.94
***WARNING*** Library was built as DEBUG. Timings may be affected.
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8        161056 ns       160886 ns         4442 TracksPropagated=397.796k/s
CPU unsync propagation/16       472188 ns       471021 ns         1490 TracksPropagated=543.5k/s
CPU unsync propagation/32      1783796 ns      1778494 ns          396 TracksPropagated=575.768k/s
CPU unsync propagation/64      6986334 ns      6961864 ns          101 TracksPropagated=588.348k/s
CPU unsync propagation/128    27832109 ns     19987935 ns           36 TracksPropagated=819.694k/s
CPU unsync propagation/256   113342481 ns     67853310 ns           10 TracksPropagated=965.848k/s
CPU sync propagation/8          165155 ns       164987 ns         4260 TracksPropagated=387.91k/s
CPU sync propagation/16         474897 ns       473924 ns         1480 TracksPropagated=540.171k/s
CPU sync propagation/32        1790464 ns      1785257 ns          393 TracksPropagated=573.587k/s
CPU sync propagation/64        7007931 ns      6971585 ns          101 TracksPropagated=587.528k/s
CPU sync propagation/128      27870228 ns     20042189 ns           35 TracksPropagated=817.476k/s
CPU sync propagation/256     114520650 ns     69456865 ns           10 TracksPropagated=943.55k/s
CUDA unsync propagation/8   4745877261 ns   4645999433 ns            1 TracksPropagated=13.7753/s
CUDA unsync propagation/16     5470499 ns      5455227 ns          118 TracksPropagated=46.9275k/s
CUDA unsync propagation/32     6119690 ns      6103531 ns          112 TracksPropagated=167.772k/s
CUDA unsync propagation/64     8467980 ns      8443357 ns           83 TracksPropagated=485.115k/s
CUDA unsync propagation/128    9593608 ns      9559909 ns           73 TracksPropagated=1.71382M/s
CUDA unsync propagation/256   19573456 ns     19390324 ns           36 TracksPropagated=3.37983M/s
CUDA sync propagation/8        5939587 ns      5925496 ns          116 TracksPropagated=10.8008k/s
CUDA sync propagation/16       5571505 ns      5555876 ns          124 TracksPropagated=46.0773k/s
CUDA sync propagation/32       5448679 ns      5434529 ns          126 TracksPropagated=188.425k/s
CUDA sync propagation/64       7596962 ns      7574358 ns           92 TracksPropagated=540.772k/s
CUDA sync propagation/128      8981539 ns      8950955 ns           76 TracksPropagated=1.83042M/s
CUDA sync propagation/256     17369342 ns     17188146 ns           41 TracksPropagated=3.81286M/s
```

after
```
Note: Google Test filter = PropagatorValidation*
[==========] Running 4 tests from 4 test suites.
[----------] Global test environment set-up.
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation1/PropagatorWithRkStepper.propagator_rk_stepper/0 (253 ms)
[----------] 1 test from PropagatorValidation1/PropagatorWithRkStepper (253 ms total)

[----------] 1 test from PropagatorValidation2/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation2/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation2/PropagatorWithRkStepper.propagator_rk_stepper/0 (290 ms)
[----------] 1 test from PropagatorValidation2/PropagatorWithRkStepper (290 ms total)

[----------] 1 test from PropagatorValidation3/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation3/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation3/PropagatorWithRkStepper.propagator_rk_stepper/0 (289 ms)
[----------] 1 test from PropagatorValidation3/PropagatorWithRkStepper (289 ms total)

[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper
[ RUN      ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0
[       OK ] PropagatorValidation4/PropagatorWithRkStepper.propagator_rk_stepper/0 (313 ms)
[----------] 1 test from PropagatorValidation4/PropagatorWithRkStepper (313 ms total)

[----------] Global test environment tear-down
[==========] 4 tests from 4 test suites ran. (1147 ms total)
[  PASSED  ] 4 tests.
 
 
 2023-06-13T15:12:20+02:00
Running ./bin/detray_array_propagator_cuda
Run on (48 X 1796.55 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 2.18, 3.90, 2.17
***WARNING*** Library was built as DEBUG. Timings may be affected.
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8        160260 ns       160133 ns         4455 TracksPropagated=399.668k/s
CPU unsync propagation/16       485918 ns       484807 ns         1445 TracksPropagated=528.045k/s
CPU unsync propagation/32      1835941 ns      1829980 ns          384 TracksPropagated=559.569k/s
CPU unsync propagation/64      7100389 ns      7079335 ns           99 TracksPropagated=578.585k/s
CPU unsync propagation/128    28557459 ns     19188468 ns           35 TracksPropagated=853.846k/s
CPU unsync propagation/256   112417637 ns     66745273 ns           10 TracksPropagated=981.882k/s
CPU sync propagation/8          162423 ns       162169 ns         4316 TracksPropagated=394.649k/s
CPU sync propagation/16         488105 ns       486809 ns         1440 TracksPropagated=525.874k/s
CPU sync propagation/32        1838921 ns      1832580 ns          382 TracksPropagated=558.775k/s
CPU sync propagation/64        7084238 ns      7061806 ns           99 TracksPropagated=580.022k/s
CPU sync propagation/128      28160175 ns     18942334 ns           37 TracksPropagated=864.941k/s
CPU sync propagation/256     112249481 ns     68144993 ns           10 TracksPropagated=961.714k/s
CUDA unsync propagation/8      6039502 ns      6024075 ns          107 TracksPropagated=10.624k/s
CUDA unsync propagation/16     5381302 ns      5364558 ns          130 TracksPropagated=47.7206k/s
CUDA unsync propagation/32     6277199 ns      6258398 ns          111 TracksPropagated=163.62k/s
CUDA unsync propagation/64     8402782 ns      8378374 ns           83 TracksPropagated=488.878k/s
CUDA unsync propagation/128    9561007 ns      9525393 ns           73 TracksPropagated=1.72003M/s
CUDA unsync propagation/256   19582703 ns     19391876 ns           36 TracksPropagated=3.37956M/s
CUDA sync propagation/8        5955242 ns      5940379 ns          114 TracksPropagated=10.7737k/s
CUDA sync propagation/16       5715400 ns      5697483 ns          121 TracksPropagated=44.9321k/s
CUDA sync propagation/32       5549028 ns      5534649 ns          124 TracksPropagated=185.016k/s
CUDA sync propagation/64       7634597 ns      7611474 ns           91 TracksPropagated=538.135k/s
CUDA sync propagation/128      8982877 ns      8951853 ns           76 TracksPropagated=1.83024M/s
CUDA sync propagation/256     17321011 ns     17134990 ns           41 TracksPropagated=3.82469M/s
```